# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HMRC API Documentation",
        "version": "1.0"
    },
    "paths": {
        "/api/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get list cars",
                "description": "get/mengambil data mobil",
                "operationId": "get/cars",
                "responses": {
                    "default": {
                        "description": "return json array data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Store cars data",
                "description": "input data mobil",
                "operationId": "cars/store",
                "requestBody": {
                    "description": "form data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "merk",
                                    "jenis",
                                    "plat",
                                    "tahun",
                                    "transmisi",
                                    "harga_weekdays",
                                    "harga_weekend"
                                ],
                                "properties": {
                                    "merk": {
                                        "type": "string"
                                    },
                                    "jenis": {
                                        "type": "string"
                                    },
                                    "plat": {
                                        "type": "string"
                                    },
                                    "tahun": {
                                        "type": "string"
                                    },
                                    "transmisi": {
                                        "type": "string"
                                    },
                                    "harga_weekdays": {
                                        "type": "string"
                                    },
                                    "harga_weekend": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return json array data"
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Delete car data",
                "description": "hapus data mobil",
                "operationId": "cars/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "insert car Id on path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return json array data"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Cars"
                ],
                "summary": "Edit cars data",
                "description": "input data mobil",
                "operationId": "cars/edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "insert car Id on path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "form data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "merk",
                                    "jenis",
                                    "plat",
                                    "tahun",
                                    "transmisi",
                                    "harga_weekdays",
                                    "harga_weekend"
                                ],
                                "properties": {
                                    "merk": {
                                        "type": "string"
                                    },
                                    "jenis": {
                                        "type": "string"
                                    },
                                    "plat": {
                                        "type": "string"
                                    },
                                    "tahun": {
                                        "type": "string"
                                    },
                                    "transmisi": {
                                        "type": "string"
                                    },
                                    "harga_weekdays": {
                                        "type": "string"
                                    },
                                    "harga_weekend": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return json array data"
                    }
                }
            }
        },
...
    }
}